slack_bot/pyproject.toml
[tool.poetry]
name = "slack-ai-bot"
version = "0.1.0"
description = ""
authors = ["Xian Wu <xianwuusa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
slack-bolt = "^1.23.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
